refactor(repo): Organizing structure with libs
